Update Tcs3472x project
